+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fix #4106: removed listing from projectrequests
 * Fix #4140: changed StatefulSet rolling pause / resume to unsupported.  Also relying on default rolling logic to Deployments and StatefulSets
 * Fix #4081: moving Versionable.withResourceVersion to a method on WatchAndWaitable and removing Waitable from the return type
+* Fix #4139: status changes don't increment metadata.generation field
 * Fix #4149: port forwarding can accept both blocking and non-blocking channels
 * Fix #4171: allowing any object in clone